requirements.txt
# Core toolkit for code analysis and LLM integration
cased-kit

# For .env file support
python-dotenv

# LLM Providers
openai
anthropic
google-generativeai

# For semantic search and embedding
sentence-transformers
chromadb

# For code execution sandbox
docker

# For schema validation (used in model_context_protocol)
jsonschema

# For static analysis
pylint
flake8
bandit
radon
lizard

# For data visualization
matplotlib
seaborn
numpy

# For git operations
GitPython

# For specific tutorials/recipes
unidiff
requests

# For dependency graph visualization (optional)
# graphviz  # Also requires Graphviz system package to be installed

# For testing
pytest
pytest-mock
pytest-cov

# FastAPI - For creating API endpoints
fastapi
uvicorn
pydantic

# Loguru - For logging.
loguru

# Watchdog - For monitoring file system events.
watchdog

# Diff Match Patch - For generating and applying patches.
diff-match-patch

# Tiktoken - For tokenizing text with OpenAI models.
tiktoken

# For syntax highlighting (often a dependency for other tools)
Pygments

# For progress bars
tqdm

# Development tools
black
mypy
ruff